--- a/benchmark/.idea/codeStyles
+++ b/benchmark/.idea/codeStyles
@@ -1,0 +1,1 @@
+../../.idea/codeStyles