--- a/benchmark/.idea/icon.png
+++ b/benchmark/.idea/icon.png
@@ -1,0 +1,1 @@
+../../.idea/icon.png