--- a/benchmark/gradle/wrapper
+++ b/benchmark/gradle/wrapper
@@ -1,0 +1,1 @@
+../../gradle/wrapper/