--- a/benchmark/gradlew.bat
+++ b/benchmark/gradlew.bat
@@ -1,0 +1,1 @@
+../gradlew.bat