--- a/benchmark/src/main/AndroidManifest.xml
+++ b/benchmark/src/main/AndroidManifest.xml
@@ -1,0 +1,2 @@
+<manifest package="com.apollographql.apollo.benchmark">
+</manifest>