--- a/composite/.idea/codeStyles
+++ b/composite/.idea/codeStyles
@@ -1,0 +1,1 @@
+../../.idea/codeStyles/