--- a/composite/.idea/icon.png
+++ b/composite/.idea/icon.png
@@ -1,0 +1,1 @@
+../../.idea/icon.png