--- a/gradle/dependencies.gradle
+++ b/gradle/dependencies.gradle
@@ -23,7 +23,6 @@ def versions = [
     rxjava3               : '3.0.6',
     rxandroid             : '2.0.1',
     sqldelight            : '1.4.3',
-    testRunner            : '0.5',
     truth                 : '1.0.1'
 ]
 
@@ -31,7 +30,6 @@ ext.dep = [
     android               : [
         plugin    : "com.android.tools.build:gradle:$versions.androidPlugin",
         minPlugin : "com.android.tools.build:gradle:$versions.minAndroidPlugin",
-        testRunner: "com.android.support.test:runner:$versions.testRunner",
     ],
     androidx              : [
         appcompat             : "androidx.appcompat:appcompat:1.1.0",