… benchmark/gradlew
… benchmark/src/androidTest/res/raw/largesample.json