Add an Android benchmark project based on [Jetpack Benchmark](https://developer.android.com/studio/profile/benchmark).

It uses the same models as [zacsweers/json-serialization-benchmarking](https://github.com/ZacSweers/json-serialization-benchmarking/tree/main/models/src/main/java/dev/zacsweers/jsonserialization/models). They're not super deep but they should give a good first point of comparison. 

Going from `2.4.1` to `3.0.0-SNAPSHOT` ([commit](https://github.com/apollographql/apollo-android/commit/8dfc59804f57aa03a11c94fe7731dc21195ed6c8))

**2.4.1 Java codegen**:
```
benchmark:    23,697,764 ns Benchmark.apollo
benchmark:    12,750,782 ns Benchmark.moshi
```

**2.4.1 Kotlin codegen**:
```
benchmark:    23,854,430 ns Benchmark.apollo
benchmark:    12,649,481 ns Benchmark.moshi
```

**3.0.0-SNAPSHOT Kotlin codegen**:
```
benchmark:    14,700,418 ns Benchmark.apollo
benchmark:    12,862,032 ns Benchmark.moshi
```

**Edit**: this was comparing 2.4.1 Java codegen to 3.0.0-SNAPSHOT Kotlin codegen. I just added the test with 2.4.1 Kotlin codegen. Let me know if I'm missing something else.

closes #2655