@AdamMTGreenberg 2021-07-02
@martinbonnin will there be any ways to hook from the client end and check specific performance benchmarks of parsing for implementation purposes, perhaps via the ApolloParseInterceptor?