@martinbonnin 2021-07-02
@AdamMTGreenberg can you elaborate a bit more? Right now I can see two scenarios:

1. Parsing benchmark (this PR): this can be made offline on test data.
2. Real User Monitoring: this requires actual production data and would also measure network speed.

Since 1. can be made offline, it can be unit tested. I don't think we need hooks in the runtime for this but maybe you have different requirements?
2. is a different story and we might need hooks to inspect the performance of the client in real life. But in this case, we'd need hooks to measure at least network, cache and parsing, maybe more.