@AdamMTGreenberg 2021-07-06
@martinbonnin thank you - our use case would involve the Real User Monitoring scenario.

Ideally, we are searching for ways to include hooks in production for network, cache and parsing as you mentioned. The goal would be to ensure that in our application in production we could test for regressions and monitor app health. The hooks/callbacks would feed into our system and allow us to trace these benchmarks. Simple start/finish would be the best.

Thanks so much