@martinbonnin 2021-07-06
@AdamMTGreenberg I'm assuming you're talking about the upcoming 3.x version ? If that's the case, let's follow up there: https://github.com/apollographql/apollo-android/issues/3221

If you're looking at 2.x, you can take a look at [`ApolloCall.Callback.onStatusEvent`](https://github.com/apollographql/apollo-android/blob/9513025a9640517f0fa448c9859938106c50606a/apollo-runtime/src/main/java/com/apollographql/apollo/ApolloCall.java#L113) that triggers Cache/Network/Completed events.